Update README.md
This Portfolio Website is a sleek, responsive site built with React and Tailwind CSS to showcase projects, skills, and contact information in a visually appealing format. Designed with a modern aesthetic, it provides an engaging experience for visitors, making it easy for them to explore showcased projects and learn about my background.